% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftconv_paired.R
\name{dcpval}
\alias{dcpval}
\title{dcpval}
\usage{
dcpval(pb, S0, right.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{pb}{A vector of Bernoulli probabilities}

\item{S0}{An observed value}

\item{right.tail}{Logical determining whether to sum the right tail (default) or left tail}

\item{log.p}{Logical determining whether to return the log of the p-value or the p-value}
}
\value{
A vector containing the log of the p-value and the p-value
}
\description{
The function takes in a vector of (Bernoulli) probabilities, an observed value s0, and calculates (log(p-value), p-value)
of the PB distribution using a direct convolution
}
\examples{
set.seed(18)
a = runif(100)
dcpval(a, 50)
dcpval(a, 50, right.tail=FALSE)
dcpval(a, 50, right.tail=TRUE, log.p=TRUE)
}
