% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftconv_paired.R
\name{fftconv}
\alias{fftconv}
\title{fftconv
Performs the 'streamlined' approach of convoluting a given probability vector in the Fourier domain,
this function prepares neccessary vectors, including the final result vector, which will be passed
to the a C program to attain the probability mass function.}
\usage{
fftconv(p)
}
\arguments{
\item{p}{Accepts a list of probabilities}
}
\value{
The final probability mass function from the convolution
}
\description{
fftconv
Performs the 'streamlined' approach of convoluting a given probability vector in the Fourier domain,
this function prepares neccessary vectors, including the final result vector, which will be passed
to the a C program to attain the probability mass function.
}
