% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftconv_paired.R
\name{logsum}
\alias{logsum}
\title{logsum
This function calculates the a sum of a vector (in log space, returns a log)}
\usage{
logsum(x)
}
\arguments{
\item{x}{A vector}
}
\description{
logsum
This function calculates the a sum of a vector (in log space, returns a log)
}
