% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftconv_paired.R
\name{shiftConvolveLog}
\alias{shiftConvolveLog}
\title{shiftConvolveLog}
\usage{
shiftConvolveLog(p, S0 = 0, process = TRUE)
}
\arguments{
\item{p}{Accepts a list of probabilities}

\item{S0}{An observed value, default set to 0}

\item{process}{Whether we want to process for 0s and 1s, always set to TRUE unless called from shiftpval}
}
\description{
Convolutes a vector of probabilities with exponential shift and returns the log of the distribution
}
\examples{
set.seed(18)
a = runif(100)
shiftConvolveLog(a, 50)
}
