% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftconv_paired.R
\name{process}
\alias{process}
\title{ShiftConvolve}
\usage{
process(p)
}
\description{
Computing the Poisson Binomial Distribution using Shifted FFT
}
\details{
A package which uses exponential shifting and Fast Fourier Transformations to
compute the distribution of the Poisson Binomial Distribution
}
\author{
Andrew Ray Lee, Noah Peres and Uri Keich
}
