% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftconv_paired.R
\name{shiftpval}
\alias{shiftpval}
\title{shiftpval}
\usage{
shiftpval(pb, S0, right.tail = TRUE)
}
\arguments{
\item{pb}{A vector of Bernoulli probabilities}

\item{S0}{An observed value}

\item{right.tail}{Logical determining whether to sum the right tail (default) or left tail}
}
\value{
A vector containing the log of the p-value and the p-value
}
\description{
The function takes in a vector of (Bernoulli) probabilities, an observed value s0, and calculates (log(p-value), p-value)
of the PB distribution using shiftConvolveLog
}
\examples{
set.seed(18)
a = runif(100)
shiftpval(a, 50)
shiftpval(a, 50, right.tail=FALSE)

}
